ui: move node to bc dir
